+++ pkg/shipper/shipper.go
@@ -38,6 +38,7 @@ type metrics struct {
 	uploads           prometheus.Counter
 	uploadFailures    prometheus.Counter
 	uploadedCompacted prometheus.Gauge
+	uploadedBytes     prometheus.Counter
 }
 
 func newMetrics(reg prometheus.Registerer, uploadCompacted bool) *metrics {
@@ -59,15 +60,14 @@ func newMetrics(reg prometheus.Registerer, uploadCompacted bool) *metrics {
 		Name: "thanos_shipper_upload_failures_total",
 		Help: "Total number of block upload failures",
 	})
-	uploadCompactedGaugeOpts := prometheus.GaugeOpts{
+	m.uploadedCompacted = promauto.With(reg).NewGauge(prometheus.GaugeOpts{
 		Name: "thanos_shipper_upload_compacted_done",
 		Help: "If 1 it means shipper uploaded all compacted blocks from the filesystem.",
-	}
-	if uploadCompacted {
-		m.uploadedCompacted = promauto.With(reg).NewGauge(uploadCompactedGaugeOpts)
-	} else {
-		m.uploadedCompacted = promauto.With(nil).NewGauge(uploadCompactedGaugeOpts)
-	}
+	})
+	m.uploadedBytes = promauto.With(reg).NewCounter(prometheus.CounterOpts{
+		Name: "thanos_shipper_uploaded_bytes_total",
+		Help: "Total number of uploaded bytes.",
+	})
 	return &m
 }
 
@@ -380,7 +380,25 @@ func (s *Shipper) upload(ctx context.Context, meta *metadata.Meta) error {
 	if err := meta.WriteToDir(s.logger, updir); err != nil {
 		return errors.Wrap(err, "write meta file")
 	}
-	return block.Upload(ctx, s.logger, s.bucket, updir, s.hashFunc)
+
+	err := block.Upload(ctx, s.logger, s.bucket, updir, s.hashFunc)
+
+	if err != nil {
+		return errors.Wrap(err, "while upploading the block")
+	}
+
+	files, err := block.GatherFileStats(updir, s.hashFunc, s.logger)
+
+	if err != nil {
+		//The block upload should not stop due to issues gathering data for a metric
+		return nil
+	}
+
+	for _, x := range files {
+		s.metrics.uploadedBytes.Add(float64(x.SizeBytes))
+	}
+
+	return nil
 }
 
 // blockMetasFromOldest returns the block meta of each block found in dir